Complete rewrite to canvas
The Game engine is completely rewritten to use canvas. There is a fallback that is not yet fully implemented using spans. It will probably be slow as all hell.

The map now moves around the player, and there are now more open things than before, but it stands on a better architecture.